Make some fixes to code phase accumulation (still unstested!).

+++ src/track.h
@@ -43,7 +43,7 @@ typedef struct {
   u32 snr_threshold_count;
   u8 prn;
 
-  u64 code_phase;
+  u32 code_phase;
   u32 code_phase_rate_fp;
 
   double code_phase_rate;
